pyproject.toml
[project]
name = "sdl_gui"
version = "0.1.0"
description = "Python SDL GUI Library"
dependencies = [
    "pysdl2>=0.9.16",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
mypy_path = "src"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
